backend/requirements.txt
Flask
requests
python-dotenv
pytz
timezonefinder
flask-cors
psycopg2-binary
SQLAlchemy